feat!: convert to typescript
Converts module to typescript, fixes some bugs that occur when
run in the browser.
