+++ .coverage/coverage-pw.json
@@ -1,0 +1,1 @@
+{"/Users/alex/Documents/Workspaces/multiformats/js-uri-to-multiaddr/src/index.ts":{"path":"/Users/alex/Documents/Workspaces/multiformats/js-uri-to-multiaddr/src/index.ts","all":false,"statementMap":{"0":{"start":{"line":1,"column":0},"end":{"line":1,"column":0}}},"s":{"0":1},"branchMap":{},"b":{},"fnMap":{},"f":{}}}
