--- .travis.yml
@@ -1,11 +1,0 @@
-language: node_js
-branches:
-  only:
-  - master
-  - /^release\/.*$/
-node_js:
-  - 'lts/*'
-  - 'node'
-script:
-  - npm run lint
-  - npm run test
