+++ package.json
@@ -1,25 +1,10 @@
 {
-  "name": "uri-to-multiaddr",
+  "name": "@multiformats/uri-to-multiaddr",
   "version": "6.0.0",
   "description": "Convert a URI to a Multiaddr: https://multiformats.io -> /dns4/multiformats.io/tcp/443/https",
-  "main": "index.js",
-  "dependencies": {
-    "is-ip": "^3.1.0",
-    "multiaddr": "^10.0.0"
-  },
-  "devDependencies": {
-    "ava": "^3.11.1",
-    "standard": "^16.0.3"
-  },
-  "engines": {
-    "node": ">=10"
-  },
-  "scripts": {
-    "lint": "standard",
-    "test": "ava test.js"
-  },
   "author": "olizilla",
-  "license": "MIT",
+  "license": "Apache-2.0 OR MIT",
+  "homepage": "https://github.com/multiformats/js-uri-to-multiaddr#readme",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/multiformats/js-uri-to-multiaddr.git"
@@ -27,5 +12,135 @@
   "bugs": {
     "url": "https://github.com/multiformats/js-uri-to-multiaddr/issues"
   },
-  "homepage": "https://github.com/multiformats/js-uri-to-multiaddr#readme"
+  "engines": {
+    "node": ">=16.0.0",
+    "npm": ">=7.0.0"
+  },
+  "type": "module",
+  "types": "./dist/src/index.d.ts",
+  "files": [
+    "src",
+    "dist/src",
+    "!dist/test",
+    "!**/*.tsbuildinfo"
+  ],
+  "exports": {
+    ".": {
+      "types": "./dist/src/index.d.ts",
+      "import": "./dist/src/index.js"
+    }
+  },
+  "eslintConfig": {
+    "extends": "ipfs",
+    "parserOptions": {
+      "sourceType": "module"
+    }
+  },
+  "release": {
+    "branches": [
+      "master"
+    ],
+    "plugins": [
+      [
+        "@semantic-release/commit-analyzer",
+        {
+          "preset": "conventionalcommits",
+          "releaseRules": [
+            {
+              "breaking": true,
+              "release": "major"
+            },
+            {
+              "revert": true,
+              "release": "patch"
+            },
+            {
+              "type": "feat",
+              "release": "minor"
+            },
+            {
+              "type": "fix",
+              "release": "patch"
+            },
+            {
+              "type": "docs",
+              "release": "patch"
+            },
+            {
+              "type": "test",
+              "release": "patch"
+            },
+            {
+              "type": "deps",
+              "release": "patch"
+            },
+            {
+              "scope": "no-release",
+              "release": false
+            }
+          ]
+        }
+      ],
+      [
+        "@semantic-release/release-notes-generator",
+        {
+          "preset": "conventionalcommits",
+          "presetConfig": {
+            "types": [
+              {
+                "type": "feat",
+                "section": "Features"
+              },
+              {
+                "type": "fix",
+                "section": "Bug Fixes"
+              },
+              {
+                "type": "chore",
+                "section": "Trivial Changes"
+              },
+              {
+                "type": "docs",
+                "section": "Documentation"
+              },
+              {
+                "type": "deps",
+                "section": "Dependencies"
+              },
+              {
+                "type": "test",
+                "section": "Tests"
+              }
+            ]
+          }
+        }
+      ],
+      "@semantic-release/changelog",
+      "@semantic-release/npm",
+      "@semantic-release/github",
+      "@semantic-release/git"
+    ]
+  },
+  "scripts": {
+    "clean": "aegir clean",
+    "lint": "aegir lint",
+    "dep-check": "aegir dep-check",
+    "generate": "protons src/pb/peer.proto src/pb/tags.proto",
+    "build": "aegir build",
+    "test": "aegir test",
+    "test:chrome": "aegir test -t browser",
+    "test:chrome-webworker": "aegir test -t webworker",
+    "test:firefox": "aegir test -t browser -- --browser firefox",
+    "test:firefox-webworker": "aegir test -t webworker -- --browser firefox",
+    "test:node": "aegir test -t node",
+    "test:electron-main": "aegir test -t electron-main",
+    "release": "aegir release"
+  },
+  "dependencies": {
+    "@multiformats/multiaddr": "^11.0.0",
+    "is-ip": "^5.0.0"
+  },
+  "devDependencies": {
+    "aegir": "^37.5.3"
+  }
 }
